+++ tests/Integration/DatasourcesApiTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace SensioLabs\Storyblok\Api\Tests\Unit;
+namespace SensioLabs\Storyblok\Api\Tests\Integration;
 
 use PHPUnit\Framework\TestCase;
 use SensioLabs\Storyblok\Api\DatasourcesApi;
